+++ imagedl/modules/sources/__init__.py
@@ -9,7 +9,10 @@ from .google import GoogleImageClient
 from ..utils import BaseModuleBuilder
 from .yandex import YandexImageClient
 from .pixabay import PixabayImageClient
+from .danbooru import DanbooruImageClient
 from .unsplash import UnsplashImageClient
+from .gelbooru import GelbooruImageClient
+from .safebooru import SafebooruImageClient
 from .duckduckgo import DuckduckgoImageClient
 
 
@@ -19,7 +22,8 @@ class ImageClientBuilder(BaseModuleBuilder):
         'BingImageClient': BingImageClient, 'BaiduImageClient': BaiduImageClient, 'GoogleImageClient': GoogleImageClient, 
         'I360ImageClient': I360ImageClient, 'PixabayImageClient': PixabayImageClient, 'YandexImageClient': YandexImageClient,
         'DuckduckgoImageClient': DuckduckgoImageClient, 'SogouImageClient': SogouImageClient, 'YahooImageClient': YahooImageClient,
-        'UnsplashImageClient': UnsplashImageClient,
+        'UnsplashImageClient': UnsplashImageClient, 'SafebooruImageClient': SafebooruImageClient, 'DanbooruImageClient': DanbooruImageClient,
+        'GelbooruImageClient': GelbooruImageClient,
     }